global object for fcci and control, rc and servo channels distibution added

+++ src/linux/cServochannels/c_servochannels.cpp
@@ -4,7 +4,7 @@ cServoChannels::cServoChannels(size_t num):
     ppm_min(900),
     ppm_max(2100)
 {
-    p_channel = new unsigned int[num];
+    p_channel = new uint16_t[num];
     for(size_t ii=0; ii<num; ii++)
         p_channel[ii] = 0;
     m_size = num;
@@ -16,18 +16,18 @@ cServoChannels & cServoChannels::operator=(cServoChannels const &other)
     {
         delete [] p_channel;
         m_size = other.m_size;
-        p_channel = new unsigned int[m_size];
-        memcpy(p_channel, other.p_channel, m_size*sizeof(unsigned int));
+        p_channel = new uint16_t[m_size];
+        memcpy(p_channel, other.p_channel, m_size*sizeof(uint16_t));
     }
     return *this;
 }
 
-void cServoChannels::setChannel(size_t no, unsigned int value)
+void cServoChannels::setChannel(size_t no, uint16_t value)
 {
     p_channel[no] = value;
 }
 
-unsigned int cServoChannels::getChannel(size_t no)
+uint16_t cServoChannels::getChannel(size_t no)
 {
     return(p_channel[no]);
 }
